+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/GravityDefying.java
@@ -1,6 +1,5 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.item.enchantment.Enchantment;
@@ -18,11 +17,6 @@ public class GravityDefying extends Enchantment implements IAttributeEnchantment
     }
 
     @Override
-    public int getMaxLevel() {
-        return 2;
-    }
-
-    @Override
     public int getMinCost(int level) {
         return 20 + (level - 1) * 30;
     }
@@ -32,24 +26,16 @@ public class GravityDefying extends Enchantment implements IAttributeEnchantment
         return this.getMinCost(level) + 20;
     }
 
-    public static final float[] GRAVITY_REDUCTION = new float[] { 0.5f, 0.7f, 0.8f };
-
-    public static float getGravityReduction(int lvl) {
-        if (lvl <= 0 || lvl > GRAVITY_REDUCTION.length)
-            return 0f;
-        return GRAVITY_REDUCTION[lvl - 1];
+    public static float getGravityReduction() {
+        return 0.55f;
     }
 
     public static void applyFallDamageReduction(LivingFallEvent event) {
-        int lvl = event.getEntity().getItemBySlot(EquipmentSlot.FEET).getEnchantmentLevel(NewEnchantmentsFeature.GRAVITY_DEFYING.get());
-        if (lvl <= 0)
-            return;
-
-        event.setDistance((event.getDistance() - lvl * 1.5f));
+        event.setDistance((event.getDistance() - 1.5f));
     }
 
     @Override
     public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
-        event.addModifier(ForgeMod.ENTITY_GRAVITY.get(), new AttributeModifier(GRAVITY_MODIFIER_UUID, "Gravity Defying enchantment", -getGravityReduction(enchantmentLvl), AttributeModifier.Operation.MULTIPLY_BASE));
+        event.addModifier(ForgeMod.ENTITY_GRAVITY.get(), new AttributeModifier(GRAVITY_MODIFIER_UUID, "Gravity Defying enchantment", -getGravityReduction(), AttributeModifier.Operation.MULTIPLY_BASE));
     }
 }